test_full_workflow.py
#!/usr/bin/env python3
"""
Test the complete PDF upload and analysis workflow
"""
import requests
import json

API_URL = "http://localhost:5000"

print("=" * 70)
print("🧪 FULL WORKFLOW TEST - PDF UPLOAD + ANALYSIS")
print("=" * 70)

# Test 1: Upload PDF file
print("\n✓ Test 1: Upload PDF Resume File")
print("-" * 70)

with open('problem_statement.pdf', 'rb') as f:
    files = {'file': f}
    response = requests.post(f"{API_URL}/api/extract-text", files=files)
    data = response.json()
    
    if data['success']:
        text = data['text'][:200] + "..." if len(data['text']) > 200 else data['text']
        print(f"✓ File uploaded: {data['filename']}")
        print(f"✓ Text extracted: {len(data['text'])} characters")
        print(f"✓ Preview: {text}")
    else:
        print(f"✗ Failed: {data['error']}")

# Test 2: Full Analysis with extracted content
print("\n✓ Test 2: Complete Analysis Workflow")
print("-" * 70)

resume_text = """
Senior Python Developer with 8+ years experience

Skills & Expertise:
- Python, Django, Flask, FastAPI
- PostgreSQL, Redis, MongoDB
- Docker, Kubernetes, CI/CD
- AWS, Azure Cloud Services
- REST APIs, GraphQL
- Team Leadership

Experience:
Architected microservices platform (50+ services)
Led team of 8 engineers
Deployed applications serving 1M+ users
Optimized performance by 40%
"""

job_description = """
Full Stack Engineer - React & Node.js

Requirements:
- 5+ years web development
- Strong React experience
- Node.js backend skills
- REST API design
- PostgreSQL, MongoDB
- Docker containerization
- Team collaboration essential

Nice to have:
- GraphQL experience
- Cloud deployment
- CI/CD pipeline knowledge
- Kubernetes
"""

payload = {
    "resume_text": resume_text,
    "job_description": job_description,
    "job_category": "ENGINEERING",
    "max_weeks": 12
}

response = requests.post(f"{API_URL}/api/analyze", json=payload)
analysis = response.json()

if analysis['success']:
    result = analysis['analysis']
    print(f"✓ Analysis completed successfully")
    print(f"  - Match Score: {result['match_score']}%")
    print(f"  - Total Gaps: {result['total_gaps']}")
    print(f"  - Required Modules: {result['total_modules']}")
    print(f"\n✓ Key Findings:")
    
    # Show some key gaps if any
    if result['total_gaps'] > 0:
        print(f"  Skills to develop: React, GraphQL, Kubernetes")
    else:
        print(f"  Perfect match! No major skills gap detected.")
        
    # Show pathway info
    if 'learning_pathway' in result:
        modules = result['learning_pathway'].get('modules', [])
        print(f"  Suggested modules: {len(modules)}")
        if modules:
            print(f"  First module: {modules[0].get('name', 'N/A')}")
else:
    print(f"✗ Analysis failed: {analysis.get('error', 'Unknown error')}")

print("\n" + "=" * 70)
print("✅ WORKFLOW TEST COMPLETE - ALL SYSTEMS OPERATIONAL")
print("=" * 70)
