requirements.txt
flask==3.0.0
werkzeug==3.0.1
numpy==1.24.3
pandas==2.0.3
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
sentence-transformers==2.2.2
transformers==4.30.2
torch==2.0.1
scipy==1.10.1
